+++ yamux/src/frame.rs
@@ -112,10 +112,8 @@ impl Frame<Data> {
         &self.body
     }
 
-    pub fn body_len(&self) -> u32 {
-        // Safe cast since we construct `Frame::<Data>`s only with
-        // `Vec<u8>` of length [0, u32::MAX] in `Frame::data` above.
-        self.body().len() as u32
+    pub fn body_len(&self) -> usize {
+        self.body().len()
     }
 
     pub fn into_body(self) -> Vec<u8> {
